+++ hole.h
@@ -8,24 +8,24 @@
 class Hole{
 private:
     std::string hole_id;
-    float hole_dip;
-    float hole_length;
+    double hole_dip;
+    double hole_length;
     coordinates hole_coordinate;
 public:
     Hole();
     ~Hole();
     void setHoleId(const char*);
-    void setHoleDip(float);
-    void setHoleLength(float);
-    void setCoordinates(float,float,float);
-    float getX();
-    float getY();
-    float getZ();
+    void setHoleDip(double);
+    void setHoleLength(double);
+    void setCoordinates(double,double,double);
+    double getX();
+    double getY();
+    double getZ();
     std::string getHoleId();
-    float getHoleDip();
-    float getHoleLength();
+    double getHoleDip();
+    double getHoleLength();
     std::vector< std::string > material_name;
-    std::vector< float > material_depth;
+    std::vector< double > material_depth;
     std::vector< std::vector< std::string > > other_material_properties;
 
 };
